roman-scribe-api/requirements.txt
fastapi
uvicorn
sqlalchemy
faster-whisper
transformers
torch
pywebview
python-multipart
pdfplumber
pydantic
playwright
fpdf2
python-dotenv
